+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod go;
+pub mod python;
 pub mod rust;
 pub mod typescript;